requirements.txt
audioread==2.1.5
baidu-aip==2.2.0.0
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
google-api-core==1.0.0
google-auth==1.4.1
google-cloud-speech==0.32.1
googleapis-common-protos==1.5.3
grpcio==1.10.0
idna==2.6
protobuf==3.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyMySQL==0.8.0
pytz==2018.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
